halite: firm things up with subtyping
The 'compatible' function was an ad hoc, intuitive attempt at
getting the desired behavior out of the collection types.
After reviewing some literature, it's clear that we actually
wish to define a subtyping relation, and ensure that the
type of a subexpression is a subtype of what is expected,
when there are expectations.

The 'meet' function replaces 'compatible', and allows
us to infer types of collection values in a well-defined way.

[This commit was extracted from another repo]