jibe: halite type-check on spec create/update and registry update
Whenever a spec is created or updated, we type check all its expressions
if any of its expressions were supplied by the user as halite.

Whenever a registry is updated, we type check all specs that
would be registered after the update if any such spec has any expression
supplied by the user as halite.

[This commit was extracted from another repo]