halite: boolean exprs to better SSA repr
We extend the SSA representation to be smarter about boolean
expressions. For each boolean node, we also add a node for its
negation and link between the two.

Moreover, we do a bit of 'normalizing' to certain operators to
so that certain equivalent expressions (e.g. (< a b), (>= b a))
actually collapse to the same node.

Having a node on hand for the negation of any boolean node
will be used in the next commit to handle computation of
'guards' that characterize when the constraints implied
by an expression node must be satisfied.

[This commit was extracted from another repo]