halite: compilation: fix+simplify inst literal elim, and add tests
We can eliminate instance literals more easily by cancelling
a get* when its immediate subform is an instance literal,
and iterating that to fixpoint. Now the pass can be
used in the presence of instance-valued variables.

We also add some test cases from previous attempts at compilation,
to see that those cases are properly handled.

[This commit was extracted from another repo]