halite: move compute-guard to j.h.t.ssa, and change signature
Analysis of guards is actually necessary to correctly restore
expressions from the SSA representation, and this isn't currently
being done right.

That problem will be fixed in a subsequent commit. To prepare, we
move compute-guards to the ssa namespace. We also change its signature,
so that we can compute guards starting from a configurable set of roots.

[This commit was extracted from another repo]