halite: various fixes that improve spec-from-ssa results
The Derivations data structure and the code that operates on it
needs shoring up. There were various errors in the code for
returning from ssa, mostly stemming from poorly defined/implemented
behaviors when rewriting nodes to produce 'alias' nodes.

This patch certainly improves things, in that observed behavior
is strictly better than it was, but there are likely still problems.

[This commit was extracted from another repo]