halite: don't rewrite unreachable negation nodes
In order to simplify the computation of guards, we ensure that
there is a node in the graph for the negation of each boolean
node. However, these negation nodes are often unreachable
from the root nodes of the constraints for a spec.

This commit modifies the graph rewriting machinery such that
unreachable negation nodes are skipped when rewriting expressions.

This produces better traces, fewer graph nodes, and will presumably
be a bit faster.

[This commit was extracted from another repo]