halite: major propagate simplification
It turns out we can construct the spec representing our bound *first*,
and then lower it along with all the others in one fell swoop.

Generating the choco spec this way allows us to delete quite a bit of
code, which was effectively just procedurally duplicating the rewrite
rules we already had.

[This commit was extracted from another repo]