halite: fix maybe/no-value related lowering rules
The rules were originally written when there was no way other
than $no-value to produce an expression of type $no-value.

Now, there are various ways (e.g. (if p $no-value (error ...))) to
produce complex forms that are statically known to evaluate
to $no-value (if they evaluate at all).

Accordingly, the rules needed updates to ensure preservation of
side effects (errors) and that they match on expressions of type
:Unset, not just the $no-value literal.

[This commit was extracted from another repo]