halite: fix massive performance issue
Recursively calling eval-expr (instead of eval-expr*) when evaluating constraints
during instance validation results in potentially exponential
redundant calls to eval-expr, in the presence of multiple levels of composition.

Before this change, an example projection was taking more than a minute
to execute. Now it's under 50ms.

The example has been added as a performance regression test.

[This commit was extracted from another repo]