choco-clj: let-bound expressions should constrain whether 'used' or not
We take the position that a variable assignment is valid iff all constraints
in the spec evaluate to a value, and further, that `let` forms evaluate
their bindings whether the resulting values are referenced or not.

Accordingly, the application of partially defined functions
(like div and expt) should have the effect of constraining the
domains of their arguments, even when they appear in let bindings
that do not contribute directly to the truth or falsity of
a constraint.

[This commit was extracted from another repo]