Improves README.me with more information
These updates are based upon the Lisp In Small Peices bookclub meeting
held on dec 10 during which Chris presented Halite to the group.

Unfortunately the current "overview" documentation link is dead.
I tried running the `(generate-local-docs)` task, but it fails due to
missing files. If you have the files, please consider adding them to the
repo so that the overview can be generated.